Add a cube debug mode
ISO level does reflect the cube debug mode when using an iso level other than 0. FIX!!

+++ Assets/Scripts/Generator/Common/NoiseSettings.cs
@@ -8,7 +8,7 @@ public class NoiseSettings : MonoBehaviour
 
     [Header("Noise sampling parameters")]
     [SerializeField]
-    private int _Seed;
+    private float _Seed;
     [SerializeField]
     private float _Frequency;
     [SerializeField]
